requirements.txt
pywebview==5.1
# Added for future expansion (Internal server for better asset loading)
Flask==3.0.0
# Useful if you decide to package into a single .exe later
pyinstaller==6.3.0
